requirements.txt
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
twilio>=9.0.0
flask>=3.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
